:ok_hand: change namespaces to avoid top level errors
thanks to havenwood
