+++ bin/eztv-scraper
@@ -1,3 +1,3 @@
 #!/usr/bin/env ruby
 
-require_relative '../lib/eztv_scraper.rb'
+require_relative '../lib/eztv_scraper'
